ref(tasks): move seer-related tasks to their own folder
