+++ pyproject.toml
@@ -676,9 +676,9 @@ module = [
   "sentry.tasks.beacon",
   "sentry.tasks.codeowners.*",
   "sentry.tasks.commit_context",
-  "sentry.tasks.delete_seer_grouping_records",
   "sentry.tasks.on_demand_metrics",
   "sentry.tasks.reprocessing2",
+  "sentry.tasks.seer.delete_seer_grouping_records",
   "sentry.tasks.store",
   "sentry.tasks.unmerge",
   "sentry.taskworker.*",
@@ -888,8 +888,8 @@ module = [
   "tests.sentry.snuba.test_tasks",
   "tests.sentry.spans.grouping.*",
   "tests.sentry.tasks.integrations.*",
+  "tests.sentry.tasks.seer.test_delete_seer_grouping_records",
   "tests.sentry.tasks.test_code_owners",
-  "tests.sentry.tasks.test_delete_seer_grouping_records",
   "tests.sentry.tasks.test_on_demand_metrics",
   "tests.sentry.tempest.*",
   "tests.sentry.templatetags.*",
