+++ src/sentry/conf/server.py
@@ -922,7 +922,7 @@ TASKWORKER_IMPORTS: tuple[str, ...] = (
     "sentry.tasks.auto_remove_inbox",
     "sentry.tasks.auto_resolve_issues",
     "sentry.tasks.auto_source_code_config",
-    "sentry.tasks.autofix",
+    "sentry.tasks.seer.autofix",
     "sentry.tasks.beacon",
     "sentry.tasks.check_am2_compatibility",
     "sentry.tasks.clear_expired_resolutions",
@@ -934,7 +934,7 @@ TASKWORKER_IMPORTS: tuple[str, ...] = (
     "sentry.tasks.commit_context",
     "sentry.tasks.commits",
     "sentry.tasks.delete_pending_groups",
-    "sentry.tasks.delete_seer_grouping_records",
+    "sentry.tasks.seer.delete_seer_grouping_records",
     "sentry.tasks.digests",
     "sentry.tasks.email",
     "sentry.tasks.files",
@@ -954,7 +954,7 @@ TASKWORKER_IMPORTS: tuple[str, ...] = (
     "sentry.tasks.repository",
     "sentry.tasks.reprocessing2",
     "sentry.tasks.scim.privilege_sync",
-    "sentry.tasks.seer",
+    "sentry.tasks.seer.cleanup",
     "sentry.tasks.statistical_detectors",
     "sentry.tasks.store",
     "sentry.tasks.summaries.weekly_reports",
@@ -972,8 +972,8 @@ TASKWORKER_IMPORTS: tuple[str, ...] = (
     "sentry.workflow_engine.tasks.delayed_workflows",
     "sentry.workflow_engine.tasks.workflows",
     "sentry.workflow_engine.tasks.actions",
-    "sentry.tasks.seer_explorer_index",
-    "sentry.tasks.context_engine_index",
+    "sentry.tasks.seer.explorer_index",
+    "sentry.tasks.seer.context_engine_index",
     # Used for tests
     "sentry.taskworker.tasks.examples",
 )
