+++ src/sentry/seer/autofix/autofix.py
@@ -47,7 +47,7 @@ from sentry.services import eventstore
 from sentry.services.eventstore.models import Event, GroupEvent
 from sentry.snuba.ourlogs import OurLogs
 from sentry.snuba.referrer import Referrer
-from sentry.tasks.autofix import check_autofix_status
+from sentry.tasks.seer.autofix import check_autofix_status
 from sentry.users.models.user import User
 from sentry.users.services.user.model import RpcUser
 from sentry.utils.event_frames import EventFrame
