+++ tests/sentry/deletions/test_group.py
@@ -184,7 +184,7 @@ class DeleteGroupTest(TestCase, SnubaTestCase):
             del os.environ["_SENTRY_CLEANUP"]
 
     @mock.patch(
-        "sentry.tasks.delete_seer_grouping_records.delete_seer_grouping_records_by_hash.apply_async"
+        "sentry.tasks.seer.delete_seer_grouping_records.delete_seer_grouping_records_by_hash.apply_async"
     )
     def test_delete_groups_delete_grouping_records_by_hash(
         self, mock_delete_seer_grouping_records_by_hash_apply_async: mock.Mock
@@ -232,7 +232,7 @@ class DeleteGroupTest(TestCase, SnubaTestCase):
         }
 
     @mock.patch(
-        "sentry.tasks.delete_seer_grouping_records.delete_seer_grouping_records_by_hash.apply_async"
+        "sentry.tasks.seer.delete_seer_grouping_records.delete_seer_grouping_records_by_hash.apply_async"
     )
     def test_invalid_group_type_handling(
         self, mock_delete_seer_grouping_records_by_hash_apply_async: mock.Mock
