+++ tests/sentry/integrations/api/endpoints/test_organization_repository_details.py
@@ -272,7 +272,7 @@ class OrganizationRepositoryDeleteTest(APITestCase):
             organization_id=org.id, key=build_pending_deletion_key(repo)
         ).exists()
 
-    @patch("sentry.tasks.seer.cleanup_seer_repository_preferences.apply_async")
+    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
     def test_put_hide_repo(self, mock_cleanup_task: MagicMock) -> None:
         self.login_as(user=self.user)
 
@@ -303,7 +303,7 @@ class OrganizationRepositoryDeleteTest(APITestCase):
             }
         )
 
-    @patch("sentry.tasks.seer.cleanup_seer_repository_preferences.apply_async")
+    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
     def test_put_hide_repo_with_commits(self, mock_cleanup_task: MagicMock) -> None:
         self.login_as(user=self.user)
 
@@ -364,7 +364,7 @@ class OrganizationRepositoryDeleteTest(APITestCase):
         assert response.data == {"integrationId": ["A valid integer is required."]}
         assert Repository.objects.get(id=repo.id).name == "example"
 
-    @patch("sentry.tasks.seer.cleanup_seer_repository_preferences.apply_async")
+    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
     def test_put_hide_repo_triggers_cleanup(self, mock_cleanup_task: MagicMock) -> None:
         """Test that hiding a repository triggers Seer cleanup task."""
         self.login_as(user=self.user)
@@ -395,7 +395,7 @@ class OrganizationRepositoryDeleteTest(APITestCase):
             }
         )
 
-    @patch("sentry.tasks.seer.cleanup_seer_repository_preferences.apply_async")
+    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
     def test_put_hide_repo_no_cleanup_when_null_fields(self, mock_cleanup_task: MagicMock) -> None:
         """Test that hiding a repository with null external_id/provider does not trigger Seer cleanup."""
         self.login_as(user=self.user)
@@ -420,7 +420,7 @@ class OrganizationRepositoryDeleteTest(APITestCase):
         # Verify the cleanup task was NOT called
         mock_cleanup_task.assert_not_called()
 
-    @patch("sentry.tasks.seer.cleanup_seer_repository_preferences.apply_async")
+    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
     def test_put_hide_repo_no_cleanup_when_external_id_null(
         self, mock_cleanup_task: MagicMock
     ) -> None:
@@ -447,7 +447,7 @@ class OrganizationRepositoryDeleteTest(APITestCase):
         # Verify the cleanup task was NOT called
         mock_cleanup_task.assert_not_called()
 
-    @patch("sentry.tasks.seer.cleanup_seer_repository_preferences.apply_async")
+    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
     def test_put_hide_repo_no_cleanup_when_provider_null(
         self, mock_cleanup_task: MagicMock
     ) -> None:
