+++ tests/sentry/tasks/seer/test_context_engine_index.py
@@ -3,7 +3,7 @@ from unittest import mock
 import pytest
 
 from sentry.seer.explorer.context_engine_utils import ProjectEventCounts
-from sentry.tasks.context_engine_index import (
+from sentry.tasks.seer.context_engine_index import (
     get_allowed_org_ids_context_engine_indexing,
     index_org_project_knowledge,
     schedule_context_engine_indexing_tasks,
@@ -28,7 +28,7 @@ class TestIndexOrgProjectKnowledge(TestCase):
         org_without_projects = self.create_organization()
         with override_options({"explorer.context_engine_indexing.enable": True}):
             with mock.patch(
-                "sentry.tasks.context_engine_index.get_event_counts_for_org_projects"
+                "sentry.tasks.seer.context_engine_index.get_event_counts_for_org_projects"
             ) as mock_counts:
                 index_org_project_knowledge(org_without_projects.id)
                 mock_counts.assert_not_called()
@@ -36,16 +36,16 @@ class TestIndexOrgProjectKnowledge(TestCase):
     def test_returns_early_when_no_high_volume_projects(self):
         with override_options({"explorer.context_engine_indexing.enable": True}):
             with mock.patch(
-                "sentry.tasks.context_engine_index.get_event_counts_for_org_projects",
+                "sentry.tasks.seer.context_engine_index.get_event_counts_for_org_projects",
                 return_value={},
             ):
                 with mock.patch(
-                    "sentry.tasks.context_engine_index.make_org_project_knowledge_index_request"
+                    "sentry.tasks.seer.context_engine_index.make_org_project_knowledge_index_request"
                 ) as mock_request:
                     index_org_project_knowledge(self.org.id)
                     mock_request.assert_not_called()
 
-    @mock.patch("sentry.tasks.context_engine_index.make_org_project_knowledge_index_request")
+    @mock.patch("sentry.tasks.seer.context_engine_index.make_org_project_knowledge_index_request")
     def test_calls_seer_endpoint_with_correct_payload(self, mock_request):
         mock_request.return_value.status = 200
 
@@ -55,19 +55,19 @@ class TestIndexOrgProjectKnowledge(TestCase):
 
         with override_options({"explorer.context_engine_indexing.enable": True}):
             with mock.patch(
-                "sentry.tasks.context_engine_index.get_event_counts_for_org_projects",
+                "sentry.tasks.seer.context_engine_index.get_event_counts_for_org_projects",
                 return_value=event_counts,
             ):
                 with mock.patch(
-                    "sentry.tasks.context_engine_index.get_top_transactions_for_org_projects",
+                    "sentry.tasks.seer.context_engine_index.get_top_transactions_for_org_projects",
                     return_value={self.project.id: ["GET /api/0/projects/"]},
                 ):
                     with mock.patch(
-                        "sentry.tasks.context_engine_index.get_top_span_ops_for_org_projects",
+                        "sentry.tasks.seer.context_engine_index.get_top_span_ops_for_org_projects",
                         return_value={self.project.id: [("db", "SELECT * FROM table")]},
                     ):
                         with mock.patch(
-                            "sentry.tasks.context_engine_index.get_sdk_names_for_org_projects",
+                            "sentry.tasks.seer.context_engine_index.get_sdk_names_for_org_projects",
                             return_value={self.project.id: "sentry.python"},
                         ):
                             index_org_project_knowledge(self.org.id)
@@ -88,27 +88,27 @@ class TestIndexOrgProjectKnowledge(TestCase):
         assert project_payload["top_transactions"] == ["GET /api/0/projects/"]
         assert project_payload["top_span_operations"] == [("db", "SELECT * FROM table")]
 
-    @mock.patch("sentry.tasks.context_engine_index.make_org_project_knowledge_index_request")
+    @mock.patch("sentry.tasks.seer.context_engine_index.make_org_project_knowledge_index_request")
     def test_raises_on_seer_error(self, mock_request):
         mock_request.return_value.status = 500
 
         with override_options({"explorer.context_engine_indexing.enable": True}):
             with mock.patch(
-                "sentry.tasks.context_engine_index.get_event_counts_for_org_projects",
+                "sentry.tasks.seer.context_engine_index.get_event_counts_for_org_projects",
                 return_value={
                     self.project.id: ProjectEventCounts(error_count=5000, transaction_count=2000)
                 },
             ):
                 with mock.patch(
-                    "sentry.tasks.context_engine_index.get_top_transactions_for_org_projects",
+                    "sentry.tasks.seer.context_engine_index.get_top_transactions_for_org_projects",
                     return_value={},
                 ):
                     with mock.patch(
-                        "sentry.tasks.context_engine_index.get_top_span_ops_for_org_projects",
+                        "sentry.tasks.seer.context_engine_index.get_top_span_ops_for_org_projects",
                         return_value={},
                     ):
                         with mock.patch(
-                            "sentry.tasks.context_engine_index.get_sdk_names_for_org_projects",
+                            "sentry.tasks.seer.context_engine_index.get_sdk_names_for_org_projects",
                             return_value={},
                         ):
                             with pytest.raises(Exception):
@@ -132,7 +132,7 @@ class TestGetAllowedOrgIdsContextEngineIndexing(TestCase):
 
         with freeze_time(frozen_time):
             with mock.patch(
-                "sentry.tasks.context_engine_index.features.has",
+                "sentry.tasks.seer.context_engine_index.features.has",
                 side_effect=feature_enabled_for_test_orgs,
             ):
                 eligible = get_allowed_org_ids_context_engine_indexing()
@@ -178,9 +178,11 @@ class TestGetAllowedOrgIdsContextEngineIndexing(TestCase):
 
 @django_db_all
 class TestScheduleContextEngineIndexingTasks(TestCase):
-    @mock.patch("sentry.tasks.context_engine_index.build_service_map.apply_async")
-    @mock.patch("sentry.tasks.context_engine_index.index_org_project_knowledge.apply_async")
-    @mock.patch("sentry.tasks.context_engine_index.get_allowed_org_ids_context_engine_indexing")
+    @mock.patch("sentry.tasks.seer.context_engine_index.build_service_map.apply_async")
+    @mock.patch("sentry.tasks.seer.context_engine_index.index_org_project_knowledge.apply_async")
+    @mock.patch(
+        "sentry.tasks.seer.context_engine_index.get_allowed_org_ids_context_engine_indexing"
+    )
     def test_dispatches_for_allowed_orgs(self, mock_get_orgs, mock_index, mock_build):
         org1 = self.create_organization()
         org2 = self.create_organization()
@@ -198,8 +200,8 @@ class TestScheduleContextEngineIndexingTasks(TestCase):
         dispatched_index_ids = [c[1]["args"][0] for c in mock_index.call_args_list]
         assert dispatched_index_ids == [org1.id, org2.id]
 
-    @mock.patch("sentry.tasks.context_engine_index.build_service_map.apply_async")
-    @mock.patch("sentry.tasks.context_engine_index.index_org_project_knowledge.apply_async")
+    @mock.patch("sentry.tasks.seer.context_engine_index.build_service_map.apply_async")
+    @mock.patch("sentry.tasks.seer.context_engine_index.index_org_project_knowledge.apply_async")
     def test_noop_when_no_allowed_orgs(self, mock_index, mock_build):
         with override_options({"explorer.context_engine_indexing.enable": True}):
             schedule_context_engine_indexing_tasks()
