+++ tests/sentry/tasks/test_post_process.py
@@ -2778,7 +2778,7 @@ class ProcessSimilarityTestMixin(BasePostProcessGroupMixin):
 
 
 class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_with_features(self, mock_generate_summary_and_run_automation):
         self.project.update_option("sentry:seer_scanner_automation", True)
@@ -2798,7 +2798,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
             event.group.id, trigger_path="old_seer_automation"
         )
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     def test_kick_off_seer_automation_without_org_feature(
         self, mock_generate_summary_and_run_automation
     ):
@@ -2816,7 +2816,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
 
         mock_generate_summary_and_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_without_scanner_on(
         self, mock_generate_summary_and_run_automation
@@ -2837,7 +2837,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
 
         mock_generate_summary_and_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_skips_existing_fixability_score(
         self, mock_generate_summary_and_run_automation
@@ -2862,7 +2862,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
 
         mock_generate_summary_and_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_runs_with_missing_fixability_score(
         self, mock_generate_summary_and_run_automation
@@ -2888,7 +2888,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
             group.id, trigger_path="old_seer_automation"
         )
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_skips_with_existing_fixability_score(
         self, mock_generate_summary_and_run_automation
@@ -2921,7 +2921,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
 
     @patch("sentry.seer.autofix.utils.is_seer_scanner_rate_limited")
     @patch("sentry.quotas.backend.check_seer_quota")
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_rate_limit_only_checked_after_all_other_checks_pass(
         self,
@@ -2992,7 +2992,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
         mock_is_rate_limited.assert_not_called()
         mock_generate_summary_and_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_skips_when_lock_held(
         self, mock_generate_summary_and_run_automation
@@ -3041,7 +3041,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
             event2.group.id, trigger_path="old_seer_automation"
         )
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature("organizations:gen-ai-features")
     def test_kick_off_seer_automation_with_hide_ai_features_enabled(
         self, mock_generate_summary_and_run_automation
@@ -3069,7 +3069,7 @@ class KickOffSeerAutomationTestMixin(BasePostProcessGroupMixin):
 class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
     """Tests for the triage signals V0 flow."""
 
-    @patch("sentry.tasks.autofix.generate_issue_summary_only.delay")
+    @patch("sentry.tasks.seer.autofix.generate_issue_summary_only.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_less_than_10_no_cache(
         self, mock_generate_summary_only, mock_seat_based_tier
@@ -3097,7 +3097,7 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         # Should call generate_issue_summary_only (not generate_summary_and_run_automation)
         mock_generate_summary_only.assert_called_once_with(group.id)
 
-    @patch("sentry.tasks.autofix.generate_issue_summary_only.delay")
+    @patch("sentry.tasks.seer.autofix.generate_issue_summary_only.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_less_than_10_with_cache(
         self, mock_generate_summary_only, mock_seat_based_tier
@@ -3130,7 +3130,7 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         "sentry.seer.autofix.utils.has_project_connected_repos",
         return_value=True,
     )
-    @patch("sentry.tasks.autofix.run_automation_only_task.delay")
+    @patch("sentry.tasks.seer.autofix.run_automation_only_task.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_gte_10_with_cache(
         self, mock_run_automation, mock_has_repos, mock_seat_based_tier
@@ -3177,7 +3177,7 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         "sentry.seer.autofix.utils.has_project_connected_repos",
         return_value=True,
     )
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_gte_10_no_cache(
         self,
@@ -3223,7 +3223,7 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         "sentry.seer.autofix.utils.has_project_connected_repos",
         return_value=False,
     )
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_gte_10_skips_without_connected_repos(
         self,
@@ -3262,7 +3262,7 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         # Should not call automation since no connected repos
         mock_generate_summary_and_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.run_automation_only_task.delay")
+    @patch("sentry.tasks.seer.autofix.run_automation_only_task.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_gte_10_skips_with_seer_last_triggered(
         self, mock_run_automation, mock_seat_based_tier
@@ -3306,7 +3306,7 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         # Should not call automation since seer_autofix_last_triggered is set
         mock_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.run_automation_only_task.delay")
+    @patch("sentry.tasks.seer.autofix.run_automation_only_task.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_event_count_gte_10_skips_with_existing_fixability_score(
         self, mock_run_automation, mock_seat_based_tier
@@ -3351,8 +3351,8 @@ class TriageSignalsV0TestMixin(BasePostProcessGroupMixin):
         # Should not call automation since seer_fixability_score is below MEDIUM threshold
         mock_run_automation.assert_not_called()
 
-    @patch("sentry.tasks.autofix.generate_summary_and_run_automation.delay")
-    @patch("sentry.tasks.autofix.run_automation_only_task.delay")
+    @patch("sentry.tasks.seer.autofix.generate_summary_and_run_automation.delay")
+    @patch("sentry.tasks.seer.autofix.run_automation_only_task.delay")
     @with_feature({"organizations:gen-ai-features": True})
     def test_triage_signals_skips_automation_for_old_issues(
         self,
